app/requirements.txt
fastapi
uvicorn
sqlalchemy
jinja2
python-multipart
dashscope
apscheduler
pydantic-settings
requests